Setup auth

+++ src/main/java/cist4830/unomaha/tempo/controllers/TagController.java
@@ -1,16 +1,23 @@
 package cist4830.unomaha.tempo.controllers;
 
 import cist4830.unomaha.tempo.controllers.errors.ResourceNotFoundException;
+import cist4830.unomaha.tempo.model.Goal;
 import cist4830.unomaha.tempo.model.Tag;
 import cist4830.unomaha.tempo.repository.GoalRepository;
 import cist4830.unomaha.tempo.repository.TagRepository;
 import cist4830.unomaha.tempo.repository.UserRepository;
+import cist4830.unomaha.tempo.controllers.utility.GetLoggedInUser;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
+import cist4830.unomaha.tempo.model.User;
+
+import org.springframework.security.core.context.SecurityContextHolder;
+import cist4830.unomaha.tempo.services.CustomUserPrincipal;
 
 import java.sql.Date;
+import java.util.stream.Collectors;
 
 @Controller
 @RequestMapping(value = "/tags")
@@ -41,7 +48,7 @@ public class TagController {
             , @RequestParam(name = "description") String description) {
         java.util.Date utilDate = new java.util.Date();
         String now = new Date(utilDate.getTime()).toString();
-        Tag tag = new Tag((long) 0, tagstr, description, (long) 1, now, now);
+        Tag tag = new Tag((long) 0, tagstr, description, GetLoggedInUser.getLoggedInUser().getId(), now, now);
         tagRepository.create(tag);
         return "redirect:/tags";
     }
@@ -75,8 +82,11 @@ public class TagController {
         Tag tag = tagRepository.findTagById(id).orElseThrow(() -> {
             throw new ResourceNotFoundException();
         });
+        User user = GetLoggedInUser.getLoggedInUser();
         model.addAttribute("tag", tag);
-        model.addAttribute("goals", tagRepository.getGoals(tag));
+        model.addAttribute("goals", tagRepository.getGoals(tag).stream()
+            .filter((goal) -> goal.getUserId().equals(user.getId()))
+            .collect(Collectors.toList()));
         return "tags/show";
     }
 
@@ -86,7 +96,10 @@ public class TagController {
             throw new ResourceNotFoundException();
         });
         // ^^ check if it exists first
-        tagRepository.delete(id);
+        // check if user owns
+        User user = GetLoggedInUser.getLoggedInUser();
+        if (user.getId() == tag.getUserId())
+            tagRepository.delete(id);
         return "redirect:/tags";
     }
 }
